[swig] Typemap for const view
This is a pretty ugly beast.  I had to find a way to slap const back onto the
pointer, especially to the data it's pointing to.  The typemap is not so pretty
either and the const_cast is surely not good practice.  At least we know it's
safe because the original object was const.