[swig] Experimental Numpy to Eigen
Numpy arrays can now be converted to Eigen types on input.  Even though we use
Eigen::Map to get a view of Numpy's memory we have to make a copy in the return
value of numpy::to because we cannot bind a reference of Eigen::Matrix<...> to
Eigen::Map<Eigen::Matrix<...>>.  On the other hand I am wondering whether this
is even a good idea.  If we only take a reference, who is going to clean up the
memory?  Also if we return the reference to a local variable the reference
immediately becomes dangling.  This needs some more thoughts and clever
engineering.