Remove GUI test deadlock
The problem was that the GUI is no longer the only thing that creates the cache
folder in the home directory. Therefore on first start of the GUI the version
file might be missing and a dialog is shown. However, we cannot dismiss this
dialog with QT_QPA_PLATFORM=offscreen and therefore we end up in a deadlock.